requirements.txt
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-JWT-Extended
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
uwsgi
psycopg2